adding the other two variants
